extract ParsedTransaction into dedicated component
- Move transaction rendering logic from history page to new component
- Maintain mobile-responsive design in component
